LPS-110880 Rename context as it is not only dealing with the id conversion it handles all the collection item stuff
