+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/floating-toolbar/LinkPanel.js
@@ -23,7 +23,7 @@ import {EDITABLE_TYPES} from '../../config/constants/editableTypes';
 import selectSegmentsExperienceId from '../../selectors/selectSegmentsExperienceId';
 import {useDispatch, useSelector} from '../../store/index';
 import updateEditableValues from '../../thunks/updateEditableValues';
-import {useGetFieldValue} from '../ControlsIdConverterContext';
+import {useGetFieldValue} from '../CollectionItemContext';
 import isMapped from '../fragment-content/isMapped';
 import MappingSelector from './MappingSelector';
